+++ src/commands/dev/Test.ts
@@ -4,7 +4,8 @@ import Command from '../../base/classes/Command';
 import CustomClient from '../../base/classes/CustomClient';
 import Category from '../../base/enums/Category';
 import { TFunction } from 'i18next';
-import { logger, useAssetsClient } from '../..';
+import { logger, useAssetsClient, useDeadlockClient } from '../..';
+import CommandError from '../../base/errors/CommandError';
 
 export default class Test extends Command {
   constructor(client: CustomClient) {
@@ -24,10 +25,23 @@ export default class Test extends Command {
     try {
       await interaction.deferReply();
 
-      const item = await useAssetsClient.ItemService.GetItem(1009965641);
+      const match = await useDeadlockClient.MatchService.GetMatch(36017246);
+      if (!match) throw new CommandError('No match found');
+
+      const player = match.players.find((player) => player.account_id === 250901865);
+      if (!player) throw new CommandError('Player not found in match');
+
+      const items = await Promise.all(
+        player.items.map(async (item) => {
+          const asset = await useAssetsClient.ItemService.GetItem(item.item_id);
+          return asset;
+        })
+      );
+
+      const response = items.map((item) => `${item?.name}`).join('\n');
 
       await interaction.editReply({
-        content: item?.name,
+        content: response,
       });
     } catch (error) {
       logger.error(error);